shop/requirements.txt
asgiref==3.4.1
coverage==6.0
Django==3.2.8
django-annoying==0.10.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
